Modulated Delay settings configurable; maximum feedback is 1.1x;

+++ Retrospector_H743_v2/src/SerialHandler.h
@@ -16,7 +16,7 @@ public:
 	void Handler(uint8_t* data, uint32_t length);
 
 private:
-	int16_t ParseInt(const std::string cmd, const char precedingChar, int low, int high);
+	int32_t ParseInt(const std::string cmd, const char precedingChar, int low, int high);
 	float ParseFloat(const std::string cmd, const char precedingChar, float low, float high);
 
 	// State machine for multi-stage commands
